Please read description below.
For this script, make sure you click the 'loop' checkbox. 
If you don't run it as a loop, then it will stop running when you receive the first message.

When you turn this script on, just simply run around all over Yew forest. 
Cover as much ground on a mount as you possibly can. 
You don't have to worry about pressing ctrl+shift at all.
This script will look around you within about a 20-25 tile radius, kind of like radar.
When you come within one of the quest objects, you will receive messages above your character's head.
Don't worry, only you can see the messages. They are not broadcast to anyone else. 
The script will alert you approx how far away the object is
and what direction according to the in-game compass.

Please note that this script will give false positives for 'Rock' on the ground, particularly
around the Yew Moongate as well as other places. The reason for this is that the sharpening stones are
the exactly same color as regular rocks in the game.
It will also show false positives if someone has a Rock around their house, and it 
might also give additional false positives that I am not aware of at this time. 

Please note, I included the script going all the way out to 50 tiles, but it does not initially
pick up items that far away. Sometimes it will go out to 30-40 tiles if you pass by the item
and keep going, but from an initial "ping", it will only scan 20-25 tiles away,
regardless of how large of a range you put into the 'range' area of the script. 

You can test this by changing the color to 'any' on the Whitestick and then buying a 'gnarled staff'
from a carpentry vendor, putting it on the ground, and then counting the tiles away from it that
the script will being alerting you to it, while changing the 'range' number.

I have heard a number of different theories on where the Whitestick spawns. 
I have found it a couple of times within a certain region in Yew forest,
mostly around the Yew Crypts and also going south from the Yew crypts to the road.
Other people claim it can spawn in Yew township. An acquaintance of mine swears that
it can spawn anywhere from the western mountain range all the way to the small eastern bridge,
on the road leaving Yew forest. The best method that I have found is to pick out a 
region and run all over it in a grid-like search pattern. If you don't find it then move
to the next region. It usually takes me a couple of hours to find it using this script. 
I have heard of people spending a few weeks running around using the Ctrl+Shift method to find it.
While this script makes it easier, you still have to get within a 20-25 tile radius
of the Whitewood stick to pick it up, so there is still a lot of searching to do.

PLEASE NOTE: if you run other scripts in UOSteam while you're running this one,
for ex, suppose you pass through a bunch of mobs and they cast on you,
and you have to use a heal macro or hide, then this script will stop running.
You need to make sure you remember to restart this script.
 
You can use this script for the melted blue wax, the sandalwood roots,
the sharpening stones, or the Whitewood stick. You can customize what you want
it to pickup in the first line. Note there is a space between the 1 & 50. It's not '150'.

You can't use @findobject for the items, because you won't know the serial of the items
which spawn on the ground. So, that's why we have to use the object's 'graphic' code.
You can adapt this script to work with other quest items that spawn on the ground a couple of ways.

One way is to find one of the objects first, and then use the 'Object Inspector' on it.
That will open a box showing you the item's 'graphic' code, which you can put into this script.
Another way is to get an exact idea of what in-game item that the quest item looks like.
For example, the Whitewood stick looks like a 'gnarled staff' which is purchasable from
any carpentry vendor. Then you get the 'graphic' code of that item to put into the script, using
the Object Inspector, but make sure you use 'any' for the color (including '').
 
I will be happy to code or customize scripts for you if you send a pm in game to 'Jiro'. 
I got the idea for this script by looking at other scripts online which use the N S E W alerts.